@@ -1,6 +1,6 @@
 {
   "name": "generator-nsis",
-  "version": "0.9.0",
+  "version": "0.10.0",
   "description": "Yeoman generator for NSIS scripts",
   "author": "Jan T. Sott",
   "license": "MIT",
@@ -11,45 +11,54 @@
     "installer",
     "setup"
   ],
-  "main": "generators/app",
+  "type": "module",
+  "exports": "./generators/app",
   "files": [
-    "generators"
+    "generators",
+    "LICENSE",
+    "README.md"
   ],
   "scripts": {
     "lint:ejs": "ejslint generators/**/*.ejs",
     "lint:js": "eslint generators/**/*.js --ignore-path .gitignore",
     "lint:json": "eslint ./*.json --ignore-path .gitignore",
     "lint": "npm-run-all --parallel lint:*",
+    "prepare": "husky",
     "publish": "np --no-yarn",
-    "test": "npm run lint"
+    "test": "uvu --ignore test/__helper.js"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/idleberg/generator-nsis"
   },
   "engines": {
-    "node": ">=14.0.0"
+    "node": ">=18"
   },
   "dependencies": {
-    "@nsis/language-data": "^0.9.0",
-    "@sindresorhus/slugify": "^1.1.2",
-    "globby": "^11.1.0",
-    "makensis": "^1.1.4",
-    "semver": "^7.3.8",
-    "spdx-license-list": "^6.6.0",
-    "terminal-link": "^2.1.1",
-    "update-notifier": "^5.1.0",
-    "yeoman-generator": "^5.7.0"
+    "@nsis/language-data": "^0.9.2",
+    "@sindresorhus/slugify": "^2.2.1",
+    "glob": "^10.3.10",
+    "makensis": "^2.0.8",
+    "semver": "^7.5.4",
+    "spdx-license-list": "^6.8.0",
+    "terminal-link": "^3.0.0",
+    "yeoman-generator": "^7.1.1"
   },
   "devDependencies": {
-    "ejs-lint": "^1.2.2",
-    "eslint": "^8.26.0",
+    "@lukeed/uuid": "^2.0.1",
+    "ejs-lint": "^2.0.0",
+    "eslint": "^8.56.0",
     "eslint-plugin-json": "^3.1.0",
-    "esprima": "^4.0.1",
-    "husky": "^8.0.1",
-    "lint-staged": "^13.0.3",
-    "np": "^7.6.2",
-    "npm-run-all": "^4.1.5"
+    "husky": "^9.0.7",
+    "lint-staged": "^15.2.0",
+    "mem-fs": "^4.0.0",
+    "np": "^9.2.0",
+    "npm-run-all2": "^6.1.1",
+    "tsm": "^2.3.0",
+    "uvu": "^0.5.6",
+    "yeoman-assert": "^3.1.1",
+    "yeoman-environment": "^4.3.0",
+    "yeoman-test": "^8.2.0"
   },
   "lint-staged": {
     "*.ejs": "ejslint",
